+++ app/src/main/java/org/privacymatters/safespace/lib/fileManager/FileItem.kt
@@ -1,4 +1,4 @@
-package org.privacymatters.safespace.lib
+package org.privacymatters.safespace.lib.fileManager
 
 class FileItem(
     name: String,
