+++ app/src/main/java/org/privacymatters/safespace/lib/fileManager/Utils.kt
@@ -1,6 +1,7 @@
-package org.privacymatters.safespace.lib
+package org.privacymatters.safespace.lib.fileManager
 
 import android.annotation.SuppressLint
+import org.privacymatters.safespace.lib.utils.Constants
 import java.text.SimpleDateFormat
 import java.util.Date
