+++ app/src/main/java/org/privacymatters/safespace/lib/mediaManager/MediaPlayer.kt
@@ -1,4 +1,4 @@
-package org.privacymatters.safespace.lib
+package org.privacymatters.safespace.lib.mediaManager
 
 import androidx.media3.exoplayer.ExoPlayer
