+++ app/src/main/java/org/privacymatters/safespace/lib/recyclerView/FolderClickListener.kt
@@ -1,6 +1,7 @@
-package org.privacymatters.safespace.lib
+package org.privacymatters.safespace.lib.recyclerView
 
 import android.view.View
+import org.privacymatters.safespace.lib.fileManager.FolderItem
 
 interface FolderClickListener {
     fun onFolderSelect(folderItem: FolderItem)
