+++ app/src/main/java/org/privacymatters/safespace/lib/recyclerView/ItemClickListener.kt
@@ -1,6 +1,7 @@
-package org.privacymatters.safespace.lib
+package org.privacymatters.safespace.lib.recyclerView
 
 import android.view.View
+import org.privacymatters.safespace.lib.fileManager.FileItem
 
 interface ItemClickListener {
     fun onClick(data: FileItem)
