+++ app/src/main/java/org/privacymatters/safespace/lib/utils/EncPref.kt
@@ -1,4 +1,4 @@
-package org.privacymatters.safespace.lib
+package org.privacymatters.safespace.lib.utils
 
 import android.content.Context
 import android.content.SharedPreferences
