+++ app/src/main/java/org/privacymatters/safespace/lib/utils/RootCheck.java
@@ -1,4 +1,4 @@
-package org.privacymatters.safespace.lib;
+package org.privacymatters.safespace.lib.utils;
 
 import java.io.File;
