+++ app/src/main/java/org/privacymatters/safespace/lib/utils/SetTheme.kt
@@ -1,4 +1,4 @@
-package org.privacymatters.safespace.lib
+package org.privacymatters.safespace.lib.utils
 
 import android.content.Context
 import androidx.appcompat.app.AppCompatDelegate
